Develop

+++ src/main/java/com/example/Personal/Expense/Tracker/service/PermissionService.java
@@ -4,6 +4,8 @@ package com.example.Personal.Expense.Tracker.service;
 import com.example.Personal.Expense.Tracker.dto.request.permission.PermissionRequest;
 import com.example.Personal.Expense.Tracker.dto.response.permission.PermissionResponse;
 import com.example.Personal.Expense.Tracker.entity.Permission;
+import com.example.Personal.Expense.Tracker.exeption.AppException;
+import com.example.Personal.Expense.Tracker.exeption.ErrorCode;
 import com.example.Personal.Expense.Tracker.mapper.PermissionMapper;
 import com.example.Personal.Expense.Tracker.repository.PermissionRepository;
 import lombok.AccessLevel;
@@ -30,8 +32,14 @@ public class PermissionService {
     public List<PermissionResponse> getAll(){
         var permissions = permissionRepository.findAll();
         return permissions.stream().map(permissionMapper::toPermissionResponse).toList();
+//        return permissions.stream().map(p -> permissionMapper.toPermissionResponse(p)).toList();
     }
     public boolean delete(String name){
+
+
+        if(!permissionRepository.existsById(name)){
+            throw new AppException(ErrorCode.PERMISSION_NOT_EXISTED);
+        }
         permissionRepository.deleteById(name);
         return true;
     }
